requirements.txt
numpy==1.21.0
Pillow==9.5.0
Pypubsub==4.0.3
six==1.16.0
wxPython==4.0.3
